refactor(marca): impide eliminación si está asociado a procesos

+++ src/marca/marca.service.ts
@@ -2,10 +2,14 @@ import { Injectable } from '@nestjs/common';
 import { CreateMarcaDto } from './dto/create-marca.dto';
 import { UpdateMarcaDto } from './dto/update-marca.dto';
 import { MarcaRepository } from './marca.repository';
+import { ProductoRepository } from '../producto/producto.repository';
 
 @Injectable()
 export class MarcaService {
-  constructor(private readonly repo: MarcaRepository) {}
+  constructor(
+    private readonly repo: MarcaRepository,
+    private readonly productoRepo: ProductoRepository,
+  ) {}
 
   async create(dto: CreateMarcaDto) {
 
@@ -27,7 +31,12 @@ export class MarcaService {
     return this.repo.update(id, dto);
   }
 
-  remove(id: number) {
+  async remove(id: number): Promise<void> {
+    const productos = await this.productoRepo.findAll();
+    const asoaciados = productos.filter(p => p.marca.id === id);
+
+    if (asoaciados.length > 0) throw new Error('No se puede eliminar la marca, ya tiene productos asoaciados');
+
     return this.repo.delete(id);
   }
 }
